Move UI dependent stuff out of CustomCourses model [#3971].
